[fix] 解决 CLAHE 问题

+++ CLAHE/imhistc.m
@@ -2,7 +2,7 @@ function [hist] = imhistc(tile)
 % imhistc - 统计直方图
 %
 
-im = tile(:);
+im = double(tile(:));
 hist = zeros(256, 1);
 for i = 1:length(im)
     hist(im(i)+1) = hist(im(i)+1) + 1;
